code/requirements.txt
amqp==2.6.1
aniso8601==7.0.0
appnope==0.1.0
backcall==0.2.0
billiard==3.6.3.0
boto3==1.16.36
botocore==1.19.36
celery==4.4.6
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.14
django-annoying==0.10.6
django-autocomplete-light==3.5.1
django-celery-beat==2.0.0
django-cognito-jwt==0.0.3
django-crispy-forms==1.9.2
django-environ==0.4.5
django-filter==2.3.0
django-js-asset==1.2.2
django-jsonstore==0.4.1
django-mailbox==4.8.1
django-maintenance-mode==0.14.0
django-material==1.6.7
django-model-utils==4.0.0
django-mptt==0.11.0
django-mssql-backend==2.8.1
django-notifications-hq==1.6.0
django-qr-code==2.0.1
django-simple-menu==1.2.1
django-tables2==2.3.1
django-timezone-field==4.0
django-viewflow==1.6.1
djangorestframework==3.11.0
djangorestframework-datatables==0.5.2
djangorestframework-simplejwt
et-xmlfile==1.0.1
flower==0.9.5
future==0.18.2
graphene==2.1.8
graphene-django==2.12.1
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
humanize==3.1.0
idna==2.10
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
jmespath==0.10.0
jsonfield==3.1.0
kombu==4.6.11
MarkupPy==1.14
numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.8
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
PyJWT==2.0.0
pyodbc==4.0.30
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
Rx==1.6.1
s3transfer==0.3.3
segno==1.3.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==2.0.0
tabulate==0.8.7
tornado==6.1
traitlets
Unidecode==1.1.1
urllib3==1.25.11
vine==1.3.0
wcwidth==0.2.5
whitenoise==4.1.4
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
zipp==3.4.0
